# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Omorix API
  description: API for the Omorix e-commerce platform
  version: 1.0.0
  contact:
    email: support@omorix.com
servers:
  - url: http://localhost:3000/api
    description: Production Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        statusCode:
          type: integer
    Error:
      type: object
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  message:
                    type: string
                  path:
                    type: array
                    items:
                      type: string
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /admin/products:
    post:
      summary: Create a product
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                  format: float
                discountPercentage:
                  type: integer
                stockQuantity:
                  type: integer
                  description: Always 0, stock managed by ProductVariant
                sku:
                  type: string
                slug:
                  type: string
                  description: unique slug for SEO
                isActive:
                  type: string
                  enum:
                    - active
                    - inactive
                    - archived
                    - restored
                    - draft
                  default: active
                categoryIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                variants:
                  type: array
                  items:
                    type: object
                    properties:
                      size:
                        type: string
                        nullable: true
                      color:
                        type: string
                        nullable: true
                      material:
                        type: string
                        nullable: true
                      price:
                        type: number
                        format: float
                      stockQuantity:
                        type: integer
                      sku:
                        type: string
                    required:
                      - stockQuantity
              required:
                - name
                - price
                - sku
                - categoryIds
                - variants
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/products/{id}:
    patch:
      summary: Update a product
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
                price:
                  type: number
                  format: float
                discountPercentage:
                  type: integer
                stockQuantity:
                  type: integer
                sku:
                  type: string
                isActive:
                  type: string
                  enum:
                    - active
                    - inactive
                    - archived
                    - restored
                    - draft
                  default: active
                categoryIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                variants:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      size:
                        type: string
                        nullable: true
                      color:
                        type: string
                        nullable: true
                      material:
                        type: string
                        nullable: true
                      price:
                        type: number
                        format: float
                      stockQuantity:
                        type: integer
                      sku:
                        type: string
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a product
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/products/{id}/stock:
    patch:
      summary: Update product variant stock
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: variantId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stockQuantity:
                  type: integer
              required:
                - stockQuantity
      responses:
        '200':
          description: Stock updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product or variant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/products/{id}/archive:
    patch:
      summary: Archive a product
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/products/{id}/restore:
    patch:
      summary: Restore a product
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/products/{productId}/images:
    post:
      summary: Add images to a product
      tags:
        - Admin
        - Products
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - images
      responses:
        '201':
          description: Images added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /products:
    get:
      summary: Get all products
      tags:
        - Public
        - Products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: isActive
          in: query
          schema:
            type: boolean
            default: true
        - name: minPrice
          in: query
          schema:
            type: number
            format: float
        - name: maxPrice
          in: query
          schema:
            type: number
            format: float
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt_desc, price_asc, price_desc]
            default: createdAt_desc
      responses:
        '200':
          description: Products retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Products not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /products/{id}:
    get:
      summary: Get product details
      tags:
        - Public
        - Products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /products/category/{id}:
    get:
      summary: Get products by category
      tags:
        - Public
        - Products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: isActive
          in: query
          schema:
            type: boolean
            default: true
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt_desc, price_asc, price_desc]
            default: createdAt_desc
      responses:
        '200':
          description: Products by category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Products or category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /products/images/{imageId}:
    get:
      summary: Get product image
      tags:
        - Public
        - Products
      parameters:
        - name: imageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Image retrieved successfully
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...